fix: display menu on mobile
